+++ firefoxThemes/modules/utils/states.js
@@ -1,12 +1,13 @@
 //Enums for the different Mixed option states
 const mixedStates = {
-  NONE: 0,
-  INITIAL: 1,
-  RESET: 2
+  NONE: 0, INITIAL: 1, RESET: 2
 };
 /*Enums for background types*/
 const backgroundTypes = {
-  PRIMARY: 0,
-  SECONDARY: 1
+  PRIMARY: 0, SECONDARY: 1
 };
-export {mixedStates,backgroundTypes};
+/*Enums for system themes states*/
+const systemStates = {
+  SYSTEM: 'system', DRUTHERS: 'druthers', DARK: 'dark', LIGHT: 'light', DEFAULT: 'default'
+};
+export {mixedStates, backgroundTypes, systemStates};